+++ api/user_api.py
@@ -2,8 +2,13 @@ from flask import *
 from flask import Blueprint
 from flask import make_response
 from flask_bcrypt import Bcrypt
+from model.auth import auth
 import jwt
-import model.database
+import os
+from dotenv import load_dotenv
+
+load_dotenv()
+jwt_secret=os.getenv("jwt_secret")
  
 bcrypt = Bcrypt()
 
@@ -20,25 +25,14 @@ def api_user():
 	password = user_content["password"]
 	hash_password = bcrypt.generate_password_hash(password).decode("utf-8")
 	try:
-		connection_object = model.database.dbconnect().get_connection()
-		cursor = connection_object.cursor(dictionary=True)
-		email_select = "SELECT * FROM user WHERE email=%s"
-		value = (email,)
-		cursor.execute(email_select,value)
-		result = cursor.fetchone()
+		result=auth.get_user(email)
 		if result != None:
 			return make_response(jsonify({"error":True,"message":"此帳號重複註冊"}),400)
 		else:
-			user = "INSERT INTO user(name,email,password) VALUES (%s,%s,%s)"
-			value = (name,email,hash_password)
-			cursor.execute(user,value)
-			connection_object.commit()
-			return make_response(jsonify({"ok":True}),200)
+			if auth.insert_user(name,email,hash_password):
+				return make_response(jsonify({"ok":True}),200)
 	except:
 		return make_response(jsonify({"error":True,"message":"伺服器錯誤"}),500)
-	finally:
-		cursor.close()
-		connection_object.close()
 
 @user_api.route("/api/user/auth",methods=["GET","PUT","DELETE"])
 def api_user_auth():
@@ -55,23 +49,15 @@ def api_user_auth():
 			user_content = request.get_json()
 			email = user_content["email"]
 			put_password = user_content["password"]
-			connection_object = model.database.dbconnect().get_connection()
-			cursor = connection_object.cursor(dictionary=True)
-			user_select = "SELECT id,name,email FROM user WHERE email=%s"
-			value = (email,)
-			cursor.execute(user_select,value)
-			user = cursor.fetchone()
+			user = auth.signin(email)
 			if user == None:
 				return make_response(jsonify({"error":True,"message":"沒有此帳號，請重新輸入"}),400)
 			else:
-				password_select = "SELECT password FROM user WHERE email=%s"
-				value = (email,)
-				cursor.execute(password_select,value)
-				password = cursor.fetchone()
+				password=auth.signin_check(email)
 				hash_password = password["password"]
 				check_password = bcrypt.check_password_hash(hash_password, put_password)
 				if check_password == True:
-					token = jwt.encode(user, "secret", algorithm="HS256")
+					token = jwt.encode(user, jwt_secret, algorithm="HS256")
 					res = make_response(jsonify({"ok":True}),200)
 					res.set_cookie('token',token,max_age=604800)
 					return res
@@ -79,9 +65,6 @@ def api_user_auth():
 					return make_response(jsonify({"error":True,"message":"密碼輸入錯誤"}),400)
 		except:
 			return make_response(jsonify({"error":True,"message":"伺服器錯誤"}),500)
-		finally:
-			cursor.close()
-			connection_object.close()
 
 	if request.method=="DELETE":
 		res = make_response(jsonify({"ok":True}),200)
