+++ app.py
@@ -1,7 +1,6 @@
 from flask import *
-from api.attractions import attractions_api
-from api.categories import categories_api
-from api.user import user_api
+from api.attractions_api import attractions_api
+from api.user_api import user_api
 
 app=Flask(
     __name__,
@@ -16,9 +15,6 @@ app.config["TEMPLATES_AUTO_RELOAD"]=True
 @app.route("/")
 def index():
 	return render_template("index.html")
-@app.route("/attraction/<id>")
-def attraction(id):
-	return render_template("attraction.html")
 @app.route("/booking")
 def booking():
 	return render_template("booking.html")
@@ -27,7 +23,6 @@ def thankyou():
 	return render_template("thankyou.html")
 
 app.register_blueprint(attractions_api, url_prefix='')
-app.register_blueprint(categories_api, url_prefix='')
 app.register_blueprint(user_api, url_prefix='')
 
 app.run(host="0.0.0.0",port=3000,debug=True)
