+++ templates/index.html
@@ -213,7 +213,8 @@
         {% endblock %}
 
         <div class="footer">COPYRIGHT © 2022 台北一日遊</div>
-
+        <script type="text/javascript" src="/js/common.js"></script>
+        
         {% block script %}
         <script type="text/javascript" src="/js/index.js"></script>
         {% endblock %}
