+++ templates/thankyou.html
@@ -1,0 +1,14 @@
+{% extends 'index.html' %}
+
+{% block title %}
+<title>台北一日遊</title>
+<link rel="stylesheet" type="text/css" href="/css/thankyou.css">
+{% endblock %}
+
+{% block content %}
+
+{% endblock %}
+
+{% block script %}
+<script type="text/javascript" src="/js/thankyou.js"></script>
+{% endblock %}
